Removes delegate-jobs.tests Replaces delegate.test.js to delegateService.test.js Adds delegateUtils.test Moves generateNotificationList test from delegate-jobs to notification-utils.test
